Removed the non-"Err"-variants of the various parallel, speculative, and sequential functions, and renamed the "Err"-variants by dropping the "Err" prefix in their names.
This is a significant API change, which simplifies the API considerably, and should not create
too much overhead in moving existing code to the new API. In our tests, the additional
check for errors does not noticeably cost performance, but is more in line with Go's error handling
practices.

If this creates any issues, please contact us.